[mac_build] separate builds, store stuff

+++ scripts/cp_artifacts.sh
@@ -4,11 +4,17 @@ set -eu
 VERSION=`cat package.json | jq -r '.version'`
 SRC=$1
 DEST=$2
-
-cp "${SRC}/mattermost-desktop-${VERSION}-win-ia32.zip" "${DEST}/mattermost-desktop-${VERSION}-win32.zip"
-cp "${SRC}/mattermost-desktop-${VERSION}-win-x64.zip" "${DEST}/mattermost-desktop-${VERSION}-win64.zip"
-cp "${SRC}/mattermost-desktop-setup-${VERSION}-win.exe" "${DEST}/"
+if [[ -f "${SRC}/mattermost-desktop-${VERSION}-win-ia32.zip" ]]; then
+    cp "${SRC}/mattermost-desktop-${VERSION}-win-ia32.zip" "${DEST}/mattermost-desktop-${VERSION}-win32.zip"
+fi
+if [[ -f "${SRC}/mattermost-desktop-${VERSION}-win-x64.zip" ]]; then
+    cp "${SRC}/mattermost-desktop-${VERSION}-win-x64.zip" "${DEST}/mattermost-desktop-${VERSION}-win64.zip"
+fi
+if [[ -f "${SRC}/mattermost-desktop-setup-${VERSION}-win.exe" ]]; then
+    cp "${SRC}/mattermost-desktop-setup-${VERSION}-win.exe" "${DEST}/"
+fi
 cp "${SRC}"/mattermost-desktop-*-mac.* "${DEST}/"
 cp "${SRC}"/mattermost-desktop-*-linux-* "${DEST}/"
 cp "${SRC}"/*.yml "${DEST}/"
 cp "${SRC}"/*.blockmap "${DEST}/"
+cp "${SRC}"/mac/*.app "${DEST}/"
